refactor: StringParser 패키지 변경
