+++ src/main/java/view/InputView.java
@@ -3,8 +3,6 @@ package view;
 import java.util.List;
 import java.util.Scanner;
 
-import domain.StringParser;
-
 public class InputView {
 
     private static final String READ_NAMES_MESSAGE = "참여할 사람 이름을 입력하세요. (이름은 쉼표(,)로 구분하세요)";
