+++ src/main/java/view/StringParser.java
@@ -1,4 +1,4 @@
-package domain;
+package view;
 
 import java.util.Arrays;
 import java.util.List;
