+++ src/test/java/view/StringParserTest.java
@@ -1,4 +1,4 @@
-package domain;
+package view;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
